Implementacion empresa

+++ src/main/java/modelo/configEmpresa/ConfiguracionEmpresa.java
@@ -1,8 +1,22 @@
 package modelo.configEmpresa;
 
+import enums.EstadoMozos;
 import exceptions.*;
+import exceptions.gestorEmpresa.EmpresaAbiertaException;
+import exceptions.mesas.MesaNoEncontradaException;
+import exceptions.mesas.MesaYaExistenteException;
+import exceptions.mesas.MesaYaOcupadaException;
+import exceptions.mozos.MozoNoEncontradoException;
+import exceptions.mozos.MozoYaAgregadoException;
+import exceptions.operarios.*;
+import exceptions.persistencia.ArchivoNoInciliazadoException;
+import exceptions.productos.ProductoEnPedidoException;
+import exceptions.productos.ProductoNoEncontradoException;
+import exceptions.productos.ProductoYaExistenteException;
+import modelo.gestorEmpresa.GestorEmpresa;
 
 import java.io.IOException;
+import java.io.Serializable;
 import java.util.ArrayList;
 
 public class ConfiguracionEmpresa {
@@ -32,40 +46,6 @@ public class ConfiguracionEmpresa {
         return instance;
     }
 
-    //NOMBRE LOCAL
-
-    /**
-     * Se encarga de cambiar el nombre del local, para esto el usuario debe ser admin,
-     * en caso de no serlo debe emitir una excepcion
-     * @param name : Nuevo nombre del local
-     * @param user : usuario que intenta realizar la accion
-     * @throws UsuarioNoAutorizadoException : Si el usuario no esta autorizado
-     * pre: El nombre debe ser un string distinto de nulo y de vacio.
-     *      user != null
-     * post: nombreLocal = name || new UsuarioNoAutorizadoException
-     *
-     */
-    public void cambiaNombreLocal(String name, Operario user) throws UsuarioNoAutorizadoException {
-        assert name != null && !name.isBlank() && !name.isEmpty() : "El nombre debe ser distinto de nulo, blanco y vacio";
-        assert user != null : "El usuario no puede ser nulo";
-
-        if(!user.puedeModificarNombreLocal())
-            throw new UsuarioNoAutorizadoException();
-        this.nombreLocal = nombreLocal;
-
-        assert this.nombreLocal == nombreLocal : "No se asigno correctamente el nombre del local";
-
-    }
-
-    /**
-     * Retorna el nombre del local
-     * @return Nombre del local
-     */
-    public String getNombreLocal(){
-        return this.nombreLocal;
-    }
-
-
     //PERSISTENCIA
 
     /**
@@ -94,6 +74,39 @@ public class ConfiguracionEmpresa {
         this.nombreLocal = pers.getNombreLocal();
         this.sueldo = pers.getSueldo();
 
+    }
+
+
+    //NOMBRE LOCAL
+
+    /**
+     * Se encarga de cambiar el nombre del local, para esto el usuario debe ser admin,
+     * en caso de no serlo debe emitir una excepcion
+     * @param name : Nuevo nombre del local
+     * @param user : usuario que intenta realizar la accion
+     * @throws UsuarioNoAutorizadoException : Si el usuario no esta autorizado
+     * pre: El nombre debe ser un string distinto de nulo y de vacio.
+     * post: nombreLocal = name || new UsuarioNoAutorizadoException
+     *
+     */
+    public void cambiaNombreLocal(String name, Operario user) throws UsuarioNoAutorizadoException, UsuarioNoLogueadoException {
+        assert name != null && !name.isBlank() && !name.isEmpty() : "El nombre debe ser distinto de nulo, blanco y vacio";
+        assert user != null : "El usuario no puede ser nulo";
+
+        if(!user.puedeModificarNombreLocal())
+            throw new UsuarioNoAutorizadoException();
+        this.nombreLocal = nombreLocal;
+
+        assert this.nombreLocal == nombreLocal : "No se asigno correctamente el nombre del local";
+
+    }
+
+    /**
+     * Retorna el nombre del local
+     * @return Nombre del local
+     */
+    public String getNombreLocal(){
+        return this.nombreLocal;
     }
 
     //SUELDO
@@ -153,8 +166,9 @@ public class ConfiguracionEmpresa {
      *      user != null
      * post: se añadira un nuevo mozo a la coleccion
      */
-    public void agregaMozo(Mozo nuevoMozo, Operario user) throws UsuarioNoAutorizadoException, MozoYaAgregadoException {
-        mozos.agregaMozo(nuevoMozo,user);
+    public void agregaMozo(Mozo nuevoMozo, Operario user) throws UsuarioNoAutorizadoException, MozoYaAgregadoException, EmpresaAbiertaException {
+        if(GestorEmpresa.getInstance().puedeAgregarMozo())
+            mozos.agregaMozo(nuevoMozo,user);
     }
 
     /**
@@ -182,9 +196,27 @@ public class ConfiguracionEmpresa {
      * pre: user != null
      * post: Se eliminara el mozo con el id ingresado de la coleccion
      */
-    public void eliminaMozo(int mozoId, Operario user) throws UsuarioNoAutorizadoException, IdIncorrectoException, MozoNoEncontradoException {
-        mozos.eliminaMozo(mozoId, user);
+    public void eliminaMozo(int mozoId, Operario user) throws UsuarioNoAutorizadoException, IdIncorrectoException, MozoNoEncontradoException, EmpresaAbiertaException {
+        if(GestorEmpresa.getInstance().puedeEliminarMozo(mozoId))
+            mozos.eliminaMozo(mozoId, user);
     };
+
+    /**
+     * Cambia el estado de un mozo
+     * pre: estado != null
+     * @param mozoId Id del mozo
+     * @param estado Estado del mozo
+     * @throws IdIncorrectoException : Si el id es incorrecto
+     * @throws MozoNoEncontradoException : Si el mozo no es encontrado;
+     */
+    public void cambiarEstadoMozo(int mozoId, EstadoMozos estado) throws IdIncorrectoException, MozoNoEncontradoException, EmpresaAbiertaException {
+        if(GestorEmpresa.getInstance().puedeDefinirEstadoMozo())
+            mozos.cambiarEstadoMozo(mozoId, estado);
+    }
+
+    public void clearEstadoMozos(){
+        mozos.clearEstadoMozos();
+    }
 
     //MESAS
 
@@ -246,8 +278,9 @@ public class ConfiguracionEmpresa {
      * pre: user != null
      * post: Se elimina la mesa de la coleccion
      */
-    public void eliminarMesa(int nroMesa, Operario user) throws UsuarioNoAutorizadoException, IdIncorrectoException, MesaNoEncontradaException {
-        mesas.eliminarMesa(nroMesa, user);
+    public void eliminarMesa(int nroMesa, Operario user) throws UsuarioNoAutorizadoException, IdIncorrectoException, MesaNoEncontradaException, MesaYaOcupadaException {
+        if(GestorEmpresa.getInstance().puedeEliminarMesa(nroMesa))
+            mesas.eliminarMesa(nroMesa, user);
     };
 
     //PRODUCTOS
@@ -308,8 +341,9 @@ public class ConfiguracionEmpresa {
      * pre: user != null
      * post: Se elimina el producto de la coleccion
      */
-    public void eliminarProducto(int idProducto, Operario user) throws UsuarioNoAutorizadoException, IdIncorrectoException, ProductoNoEncontradoException {
-        productos.eliminarProducto(idProducto, user);
+    public void eliminarProducto(int idProducto, Operario user) throws UsuarioNoAutorizadoException, IdIncorrectoException, ProductoNoEncontradoException, ProductoEnPedidoException {
+        if(GestorEmpresa.getInstance().puedeEliminarProducto(idProducto))
+            productos.eliminarProducto(idProducto, user);
 
     };
 
@@ -390,7 +424,7 @@ public class ConfiguracionEmpresa {
      *      password != null && password != ""
      * post: retorna el operario deseado.
      */
-    public Operario login(String nombreDeUsuario, String password) throws DatosLoginIncorrectosException {
+    public Operario login(String nombreDeUsuario, String password) throws DatosLoginIncorrectosException, OperarioInactivoException {
         return operarios.login(nombreDeUsuario,password);
     }
